Export protocols to canonical session type
This commit adds an option to export a protocol to the canonical
multiparty session type. This is done by converting the global graph to
a graph with only send send actions. This graph is basically a data
structure view of how the code looks. This graph is used to output a
session type using minimal fixpoints, variable recursion and labelled
choice in combination with a send action.

This work is a preparation for creating a tool to convert the multiparty
protocols to binary session protocols according to the method described
in
Luís Caires, and Jorge A. Pérez. "A typeful characterization of
multiparty structured conversations based on binary sessions." arXiv
preprint arXiv:1407.4242 (2014).

This commit also removes some trailing spaces.